test: extract shell CLI white-box suite
