+++ scripts/rust-source-size-baseline.txt
@@ -20,7 +20,6 @@ crates/agentfs/src/lib.rs 1151
 crates/agentfs/src/root.rs 1191
 crates/agentfs/tests/agent_files.rs 1020
 crates/agentfs/tests/agent_root_overlay.rs 1645
-crates/alan/src/cli/shell.rs 1322
 crates/alan/src/cli/skill_authoring.rs 1187
 crates/alan/src/cli/skill_eval.rs 1167
 crates/alan/src/main.rs 1370
